Refactor "held node" concept to an array of selections
Rather than thinking of a single node which is held, we now need to
support a selection of one or more nodes. Even if we're only moving a
single node as we were before, we need to think of this in the same way,
so it's now a selection of one.